surrealql: Refactor statements
